# backend/schema/openapi.yaml
openapi: 3.0.3
info:
  title: DataForge API
  version: v1
  description: Synthetic data streams for data-engineering education.
paths:
...
  /api/v1/schemas:
    get:
      operationId: schemas_list
      description: 'GET /schemas (api-spec §4.12 #62).'
      tags:
      - schemas
      security:
      - apiKeyAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectSummary'
          description: ''
  /api/v1/schemas/{subject}:
    get:
      operationId: schemas_retrieve
      description: 'GET /schemas/{subject} (api-spec §4.12 #63).'
      parameters:
      - in: path
        name: subject
        schema:
          type: string
        required: true
      tags:
      - schemas
      security:
      - apiKeyAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectDetail'
          description: ''
  /api/v1/schemas/{subject}/diff:
    get:
      operationId: schemas_diff_retrieve
      description: |-
        GET /schemas/{subject}/diff?from=a&to=b (api-spec §4.12 #66).

        The computed added-fields diff between two registered versions (§5.3): ``added``
        is ``properties(to) \ properties(from)`` recursively (incl. fields added inside
        an existing nested object), each ``{path, type, required:false}``. Under
        ``BACKWARD_ADDITIVE`` ``removed``/``changed`` are empty by construction
        (INV-REG-3); they exist in the shape for V-2. ``404`` if either version is
        absent; ``400`` ``validation-error`` if ``from ≥ to`` (a diff is forward-only).
        The diff is computed, never stored (§3.2). Auth is ``schemas:read`` (A-4).
      parameters:
      - in: query
        name: from
        schema:
          type: integer
        required: true
      - in: path
        name: subject
        schema:
          type: string
        required: true
      - in: query
        name: to
        schema:
          type: integer
        required: true
      tags:
      - schemas
      security:
      - apiKeyAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDiff'
          description: ''
  /api/v1/schemas/{subject}/versions:
    get:
      operationId: schemas_versions_list
      description: 'GET /schemas/{subject}/versions (api-spec §4.12 #64).'
      parameters:
      - in: path
        name: subject
        schema:
          type: string
        required: true
      tags:
      - schemas
      security:
      - apiKeyAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VersionProvenance'
          description: ''
  /api/v1/schemas/{subject}/versions/{schema_version}:
    get:
      operationId: schemas_version_retrieve
      description: |-
        GET /schemas/{subject}/versions/{version} (api-spec §4.12 #65).

        ``{version}`` is an integer or the literal ``latest``. The ``schema`` member is
        the stored document verbatim (annotations included) — what consumer pipelines
        fetch to resolve a ``schema_ref`` (exercise E5).
      parameters:
      - in: path
        name: schema_version
        schema:
          type: string
        required: true
      - in: path
        name: subject
        schema:
          type: string
        required: true
      tags:
      - schemas
      security:
      - apiKeyAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionRecord'
          description: ''
...
components:
  schemas:
    AddedField:
      type: object
      description: One added property in the computed diff (#66).
      properties:
        path:
          type: string
        type:
          type: string
        required:
          type: boolean
      required:
      - path
      - required
      - type
...
    SchemaDiff:
      type: object
      description: |-
        The computed version-to-version diff (#66).

        Under ``BACKWARD_ADDITIVE`` (the only MVP mode) ``removed_fields`` /
        ``changed_fields`` are always empty by construction (INV-REG-3); they are in the
        shape so the contract survives a future compatibility-mode addition (V-2).
      properties:
        subject:
          type: string
        from_version:
          type: integer
        to_version:
          type: integer
        added_fields:
          type: array
          items:
            $ref: '#/components/schemas/AddedField'
        removed_fields:
          type: array
          items:
            $ref: '#/components/schemas/AddedField'
        changed_fields:
          type: array
          items:
            $ref: '#/components/schemas/AddedField'
      required:
      - added_fields
      - changed_fields
      - from_version
      - removed_fields
      - subject
      - to_version
...
    SubjectDetail:
      type: object
      description: 'Subject detail (#63): the summary plus created_at + per-version
        provenance.'
      properties:
        subject:
          type: string
        scenario_slug:
          type: string
        compatibility:
          type: string
        latest_version:
          type: integer
          nullable: true
        versions:
          type: array
          items:
            type: integer
        created_at:
          type: string
          format: date-time
        version_provenance:
          type: array
          items:
            $ref: '#/components/schemas/VersionProvenance'
      required:
      - compatibility
      - created_at
      - latest_version
      - scenario_slug
      - subject
      - version_provenance
      - versions
    SubjectSummary:
      type: object
      description: Subject summary (#62).
      properties:
        subject:
          type: string
        scenario_slug:
          type: string
        compatibility:
          type: string
        latest_version:
          type: integer
          nullable: true
        versions:
          type: array
          items:
            type: integer
      required:
      - compatibility
      - latest_version
      - scenario_slug
      - subject
      - versions
...
    VersionProvenance:
      type: object
      description: 'One version''s provenance row (#63 detail, #64 list).'
      properties:
        version:
          type: integer
        registered_at:
          type: string
          format: date-time
        manifest_version:
          type: string
          nullable: true
      required:
      - manifest_version
      - registered_at
      - version
    VersionRecord:
      type: object
      description: One version's full record including the schema document (#65).
      properties:
        subject:
          type: string
        version:
          type: integer
        manifest_version:
          type: string
          nullable: true
        registered_at:
          type: string
          format: date-time
        schema:
          type: object
          additionalProperties: {}
      required:
      - manifest_version
      - registered_at
      - schema
      - subject
      - version
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT